UpdatePrototype
added slide to playerMovment,
playerMovement was made easy for artist and easy to tweak in inspector,
Added Enemy state feeding this will alow enemy to go into a feeding state neer a foodSource Waypoint.
Lots more work was done on door system but no solution was found as of yet.